Extract Bubblewrap sandbox implementation
